refactor(test): share mobile_sync mockall mocks via test_support
Address PR #725 review: the P1 mobile_sync slice added a hand-rolled
`CapturingAnalyticsSink` to three use case test modules and re-pasted
the same `mockall::mock!` blocks (DeviceRepo / Hasher / Minter / Staging)
that already lived in several existing siblings. Reviewer asked to stop
re-implementing per file and consolidate via mockall.

New `mobile_sync/test_support.rs` declares each shared mock exactly once
(plus `CapturingAnalyticsSink`); seven use case tests now `use` them
instead of redefining. Per-file unique mocks (Settings/Clock/Probe in
register_device) and stateful fakes (InMemorySettings, FakeEntryRepo,
FixedProbe) stay local — rationale documented at the module head.

Net -300 lines; uc-application 464 / uc-observability 83 /
uc-bootstrap 27 / uc-webserver 61 tests all green.
